Adds kube_nodes_provider starlark built-in
This patch includes the implementation of the kube_nodes_provider
starlark built-in. This provider allows listing of the hosts in a
k8s cluster accessible by the provided/default kube config.
It also includes identification of the provider in the resources
directive, so that the crashd config script can use the
kube_nodes_provider to enumerate the node IP addresses of the hosts to
run crashd commands on.