Passes script specfic args to the diagnostics file
This patch adds the ability to allow the users to pass an arbitrary map
of key value pairs while running the program via the --args flag. These
arguments are available to the crash diagnostics file as a map by the
name of args. Script designers can use this ability to design static
scripts which can be populated by different values on runtime.