New provider for CAPV resource enumeration
This patch adds a new capv_provider() which is used to enumerate the
compute resources for the management/or workload clusters managed by
CAPV.
It also extends the kube_config() directive to accept a CAPI provider
implementation as an input, thus enabling workload cluster kube configs
to be discoverable by the directive and used in the script subsequently.

Also, this patch updates the generation of sshArgs by including the
private key path.

In addition, it also updates the SearchParams struct from the k8s
package to accept the search parameters directly as golang slices.

Adds private key path to ssh args