Adds the set_as_default directive
This patch introduces a new starlark directive called as
`set_as_default` which can be used to set the default values of one of
the three types:
- kube_config
- ssh_config
- resources

Previously, on each call of kube_config, ssh_config or resources, the
resulting value was set in the local starlark thread and avialbale
inside the script. This patch removes the need for this default
behavior. Instead, the script developer needs to make a choice to call
the set_as_default directive with the value that needs to be stored in
the thread.