Includes a new provider for CAPA managed objects
Adds a new resource provider which can be used to list out resources for
a CAPA management/workload cluster.
This also includes updating the capv_provider to include the namespace
for the queried workload cluster.
It also updates the ssh_config() directive to add jump host and user
only when provided.