Updates ssh test helpers to use generated ssh keys
This patch updates the ssh test helpers to generate a random
public/private key-pair to be used for SSH server related tests. The
helpers now create a new key-pair and write them to a temporary file
location everytime the test ssh server gets boooted up.
The test helpers also randomize the genration of usernames instead of
hardcoding a static username in the tests