Changes ssh command string for proxy args
The usage of the -J flag requires the configuration options for jump
hosts to be set in the ~/.ssh/config file. It does not respect the
inline options sent vvia the command string. Hence, reverted to using
the ProxyCommand option to connect to hosts in case of the presence of
a jumpbox.