Adds args-file flag to run command
This patch adds a new flag `--args-file` to the run command which takes
the path to a file containing key=value pairs of arguments which can be
passed to the Starlark diagnostics script.
The values for the keys passed via the `--args-file` flag can be
overridden by passing a diff value for the same key via the `--args`
flag.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>