Adds support for passphrase protected ssh key
This patch adds support for an instance of ssh-agent which is used to
communicate (ssh/scp) with remote compute resources.
Whenever the Starlark crashd_config() directive uses the `use_ssh_agent`
parameter, a new instance of ssh.Agent is started and keys are added to
this Agent instance. All ssh and scp operations then use this agent to
connect to remote resources.

This patch also includes:
- some variable name changes for some Starlark builtins.
- bumping of the test workflow to use go 1.15 since we use rely on new
  functionality introduced post go 1.14

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>