Enhance Kubernetes search and avoid overwriting captured object logs
The main implementation in this patch is the simplification of the search
algorithm to better apply group, versions, resources relationships even
when server contains resources of same groups, but different versions.

Another fix is the overwriting of captured files during Kubernetes search.
This fix append a timestamp at the end of each file, it also adds the
group and version of the resource to the path where it is saved.

Also, the patch updates the search tests and converted some of the ginkgo
driven tests back to simple Go test functions.

Signed-off-by: Vladimir Vivien <vivienv@vmware.com>